backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
pycryptodome==3.20.0
pydantic==2.6.4
python-dotenv==1.0.1
supabase==2.4.5
